--- a/extern/detray/CMakeLists.txt
+++ b/extern/detray/CMakeLists.txt
@@ -13,7 +13,7 @@ message( STATUS "Building Detray as part of the TRACCC project" )
 
 # Declare where to get Detray from.
 set( TRACCC_DETRAY_SOURCE
-   "URL;https://github.com/acts-project/detray/archive/refs/tags/v0.93.0.tar.gz;URL_MD5;c01cc864316b65bc8c88639b709e9816"
+   "URL;https://github.com/acts-project/detray/archive/refs/tags/v0.94.0.tar.gz;URL_MD5;b9f6c8f0fa5c6df22df1ebe1aff2ab4a"
    CACHE STRING "Source for Detray, when built as part of this project" )
 mark_as_advanced( TRACCC_DETRAY_SOURCE )
 FetchContent_Declare( Detray SYSTEM ${TRACCC_DETRAY_SOURCE} )