Some benchmark number will be added later

EDIT: There is a slight performance improvement

### ODD 

#### This PR

```
12:37:40    CudaSeqExample                INFO      ==> Statistics ... 
12:37:40    CudaSeqExample                INFO      - read    3790687 cells
12:37:40    CudaSeqExample                INFO      - created (cpu)  0 measurements     
12:37:40    CudaSeqExample                INFO      - created (cuda)  1042450 measurements     
12:37:40    CudaSeqExample                INFO      - created (cpu)  0 spacepoints     
12:37:40    CudaSeqExample                INFO      - created (cuda) 823334 spacepoints     
12:37:40    CudaSeqExample                INFO      - created  (cpu) 0 seeds
12:37:40    CudaSeqExample                INFO      - created (cuda) 368854 seeds
12:37:40    CudaSeqExample                INFO      - found (cpu)    0 tracks
12:37:40    CudaSeqExample                INFO      - found (cuda)   918911 tracks
12:37:40    CudaSeqExample                INFO      - fitted (cpu)   0 tracks
12:37:40    CudaSeqExample                INFO      - fitted (cuda)  918911 tracks
12:37:40    CudaSeqExample                INFO      ==>Elapsed times...            File reading  (cpu)  7935 ms
12:37:40    CudaSeqExample                INFO               Clusterization (cuda)  32 ms
12:37:40    CudaSeqExample                INFO         Spacepoint formation (cuda)  4 ms
12:37:40    CudaSeqExample                INFO                      Seeding (cuda)  378 ms
12:37:40    CudaSeqExample                INFO                 Track params (cuda)  4 ms
12:37:40    CudaSeqExample                INFO                Track finding (cuda)  2894 ms
12:37:40    CudaSeqExample                INFO                Track fitting (cuda)  4088 ms
12:37:40    CudaSeqExample                INFO                           Wall time  15410 ms
```

#### main 476f725

```
11:53:15    CudaSeqExample                INFO      ==> Statistics ... 
11:53:15    CudaSeqExample                INFO      - read    3790687 cells
11:53:15    CudaSeqExample                INFO      - created (cpu)  0 measurements     
11:53:15    CudaSeqExample                INFO      - created (cuda)  1042450 measurements     
11:53:15    CudaSeqExample                INFO      - created (cpu)  0 spacepoints     
11:53:15    CudaSeqExample                INFO      - created (cuda) 823334 spacepoints     
11:53:15    CudaSeqExample                INFO      - created  (cpu) 0 seeds
11:53:15    CudaSeqExample                INFO      - created (cuda) 368854 seeds
11:53:15    CudaSeqExample                INFO      - found (cpu)    0 tracks
11:53:15    CudaSeqExample                INFO      - found (cuda)   918924 tracks
11:53:15    CudaSeqExample                INFO      - fitted (cpu)   0 tracks
11:53:15    CudaSeqExample                INFO      - fitted (cuda)  918924 tracks
11:53:15    CudaSeqExample                INFO      ==>Elapsed times...            File reading  (cpu)  7917 ms
11:53:15    CudaSeqExample                INFO               Clusterization (cuda)  33 ms
11:53:15    CudaSeqExample                INFO         Spacepoint formation (cuda)  4 ms
11:53:15    CudaSeqExample                INFO                      Seeding (cuda)  378 ms
11:53:15    CudaSeqExample                INFO                 Track params (cuda)  4 ms
11:53:15    CudaSeqExample                INFO                Track finding (cuda)  2935 ms
11:53:15    CudaSeqExample                INFO                Track fitting (cuda)  4157 ms
11:53:15    CudaSeqExample                INFO                           Wall time  15474 ms
```

### Continuous benchmark with toy detector

#### This PR

```
Detector check: OK
----------------------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------
ToyDetectorBenchmark/CUDA/real_time 6926930398 ns   6926525294 ns            1 event_throughput_Hz=14.4364/s
```

#### main 476f725

```
Detector check: OK
----------------------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------
ToyDetectorBenchmark/CUDA/real_time 6952780753 ns   6952405253 ns            1 event_throughput_Hz=14.3827/s
```